# openapi.yaml
openapi: 3.0.3
info:
  title: Controle Saudável API
  version: 1.0.0
servers:
  - url: http://localhost:5001
paths:
  /auth/signup:
    post:
      summary: Cadastro de usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string }
                password: { type: string }
                age: { type: integer }
                height: { type: number }
                weight: { type: number }
                sex: { type: string }
                goal: { type: string }
      responses:
        '201':
          description: Usuário criado
  /food/search:
    get:
      summary: Busca de alimentos
      parameters:
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200':
          description: Lista de alimentos
  /image/recognize:
    post:
      summary: Reconhecimento de imagem de refeição
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Estimativa nutricional
  /user/{id}/meals:
    get:
      summary: Lista refeições do usuário
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: date
          schema: { type: string, format: date }
      responses:
        '200':
          description: Lista de refeições
    post:
      summary: Cria refeição para usuário
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date: { type: string, format: date }
                items:
                  type: array
                  items:
                    type: object
                totalCalories: { type: number }
                totalProtein: { type: number }
                totalCarbs: { type: number }
                totalFat: { type: number }
      responses:
        '201':
          description: Refeição criada
  /user/{id}/profile:
    put:
      summary: Atualiza perfil do usuário
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Perfil atualizado